# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Digital HubX API.",
        "title": "Digital HubX API",
        "contact": {
            "name": "API Support (Tanapong R)",
            "url": "mailto:zTanapongR@pttep.com",
            "email": "zTanapongR@pttep.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/hubx-service",
    "paths": {
        "/admin/attendance": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to create an attendance record for a specific class session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Attendance"
                ],
                "summary": "Create a new attendance record",
                "parameters": [
                    {
                        "description": "Create Attendance Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAttendanceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"Attendance created\",\"status\":200,\"details\":{\"field\":\"value\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/choice": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to create a choice for a specific question.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Choice"
                ],
                "summary": "Create a new choice",
                "parameters": [
                    {
                        "description": "Create Choice Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateChoiceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"Choice created\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateAttendanceReq": {
            "description": "Request body for creating an attendance record.",
            "type": "object",
            "required": [
                "class_session_id",
                "user_email"
            ],
            "properties": {
                "class_id": {
                    "description": "Class ID",
                    "type": "string"
                },
                "class_session_id": {
                    "description": "Class session ID",
                    "type": "string"
                },
                "user_email": {
                    "description": "User's email",
                    "type": "string"
                }
            }
        },
        "dto.CreateChoiceReq": {
            "description": "Request body for creating a new choice.",
            "type": "object",
            "required": [
                "description",
                "question_id"
            ],
            "properties": {
                "description": {
                    "description": "Choice description",
                    "type": "string"
                },
                "question_id": {
                    "description": "Question ID",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "This security definition is used for authenticating",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}